Change Unix converge strategy to InstallSh
This change makes the unix converge strategy InstallSh.
There is a bug in provisioning in the Cached install
strategy with the new chef_version changes. InstallSh
works so changeing to it until the lower level stuff is
fixed.

Updated the tests to use vagrants insecure_key and changed
the machine file to a download so we are not reliant on local
state.