Allow key_data parameter for SSH keys, avoid overwriting command prefix
key_data added as an additional parameter for specifying SSH keys.
Additionally, the command prefix is only set if it doesn't already exist,
allowing a user to explicitly set it as an empty string or something of
their choosing when not connecting as root. This includes a different
method of avoiding the override from PR #20